move replication implementation to its own module
make it more clear which pieces are for replication and which are for recovery
also hopefully make it easier for work to proceed in parallel on these two fronts
