+++ src/_zkapauthorizer/recover.py
@@ -20,22 +20,14 @@ from typing import BinaryIO, Callable, Iterator, Optional
 
 from attrs import define
 from twisted.python.filepath import FilePath
-from twisted.python.lockfile import FilesystemLock
 
-from .config import REPLICA_RWCAP_BASENAME
 from .sql import escape_identifier
-from .tahoe import Tahoe, attenuate_writecap
+from .tahoe import Tahoe
 
 
 class SnapshotMissing(Exception):
     """
     No snapshot was not found in the replica directory.
-    """
-
-
-class ReplicationAlreadySetup(Exception):
-    """
-    An attempt was made to setup of replication but it is already set up.
     """
 
 
@@ -273,47 +265,3 @@ def get_tahoe_lafs_downloader(client: Tahoe) -> Callable[[str], Downloader]:
         return downloader
 
     return get_downloader
-
-
-async def fail_setup_replication():
-    """
-    A replication setup function that always fails.
-    """
-    raise Exception("Test not set up for replication")
-
-
-async def setup_tahoe_lafs_replication(client: Tahoe) -> Awaitable[str]:
-    """
-    Configure the ZKAPAuthorizer plugin that lives in the Tahoe-LAFS node with
-    the given configuration to replicate its state onto Tahoe-LAFS storage
-    servers using that Tahoe-LAFS node.
-    """
-    # Find the configuration path for this node's replica.
-    config_path = client.get_private_path(REPLICA_RWCAP_BASENAME)
-
-    # Take an advisory lock on the configuration path to avoid concurrency
-    # shennanigans.
-    config_lock = FilesystemLock(config_path.path + ".lock")
-    config_lock.lock()
-    try:
-
-        # Check to see if there is already configuration.
-        if config_path.exists():
-            raise ReplicationAlreadySetup()
-
-        # Create a directory with it
-        rw_cap = await client.make_directory()
-
-        # Store the resulting write-cap in the node's private directory
-        config_path.setContent(rw_cap.encode("ascii"))
-
-    finally:
-        # On success and failure, release the lock since we're done with the
-        # file for now.
-        config_lock.unlock()
-
-    # Attenuate it to a read-cap
-    rocap = attenuate_writecap(rw_cap)
-
-    # Return the read-cap
-    return rocap
